Separate app and http server files
